+++ .github/CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/), and this project *somewhat* adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2020-06-08
+### Added
+ - Added support for the following WoW Classic Game Data endpoint category :
+   - WoW Token
+     - Note : for WoW Classic, the Token is only available in China
+
 ## [1.3.0] - 2020-06-07
 ### Added
  - Added support for the following Hearthstone Game Data endpoint category :